move gas config to app
