+++ bin/src/tx.rs
@@ -3,9 +3,9 @@ use {
     anyhow::anyhow,
     clap::{Parser, Subcommand},
     colored::Colorize,
+    grug_app::GAS_COSTS,
     grug_sdk::{Client, GasOption, SigningKey, SigningOption},
     grug_types::{from_json_slice, Addr, Binary, Coins, Hash, Message, UnsignedTx},
-    grug_vm_wasm::GAS_COSTS,
     serde::Serialize,
     std::{fs::File, io::Read, path::PathBuf, str::FromStr},
     tendermint_rpc::endpoint::broadcast::tx_sync,
